Lvl3 slight fix and Tutorial revamp
I made the indicator orbs non-trigger, then, in Tutorial, I added a new pathway to add dilemma and fixed the shooting enemies so that the blastwave is based on the enemy position and rotation. They can also be killed if attacked, hitting them without attacking triggers a "your dead" print as I don't have access to player scripts. The battery now only prints when necessary in order to preserve system performance. Lastly, the door to the windmill has an override switch inside which makes the door open permanently.